bindings: introduce schema-based bindings
Add native <-> ovs format bindings based on schema

In order to be able to translate ovsdb data to and from libovsdb format,
we need to know the schema of the column. Without it we would not know
which string to convert to UUID or what's the type of empty sets.

Based on DatabaseSchema, we can infer the type of each column and make a
deterministic translation

bindings.go really offers two functions OvnToNative and NativeToOvn

Signed-off-by: Adrian Moreno <amorenoz@redhat.com>