orm: introduce ORM layer
The ORM layer is a wrapper around the native bindings that enable the
use of tagged structs instead of map[string]interface{}

The handling of the ORM objects is performaned by ormInfo structure that
provides ORM metadata and common operations

A ORM tagged struct is one that has fields tagged with the 'ovs' tag.
The value of the tag indicates the column in the table where the data is
to be written / read from.
Eg:
type myType struct {
    MyField string `ovs:"myColumn"`
}

Signed-off-by: Adrian Moreno <amorenoz@redhat.com>