cache: Implement a cache on the client
The client now includes a cache, and also registers a handler
to ensure the cache is updated once Monitor or MonitorAll is called.
It also permits clients to register callbacks for Add/Update/Delete
cache events, which is a nicer API that registering your own handler and
unpacking the updates yourself.

Signed-off-by: Dave Tucker <dave@dtucker.co.uk>