encoding_test: Compare unmarshalled set in maps
Golang's maps do not have any order guarantee when copying or iterating.
That's not usually a problem when you compare their JSON representations
because assert.JSONEq uses reflect.DeepEqual which takes this into
account

However, in this case, the unmarshalled represetation of the map is
actually a slice, whose ordering does matter.

Therefore, a more robust comparison would be to use
assert.ElementsMatch()

Signed-off-by: Adrian Moreno <amorenoz@redhat.com>