Separation of default values and unset fields in MonitorSelect object .
The RFC 7047 defines that if any of the MonitorSelect parameters is not set,
it equals to 'true'.
For example, for "Insert" - 'If "insert" is omitted or true, "update"
notifications are sent for rows newly inserted into the table.'

The current implementation uses bools for the MonitorSelect fields,
so unset fields get the default value -"false", which violates the RFC
definition.

Another small change here is a replacement of the Select field in MonitorRequest
from the MonitorSelect struct by a pointer to the struct.
The Select field has the "omitempty" tag, but the tag doesn't make sense for
structs and Json marshaling adds empty struct (Select[]) for unset Select field.

Signed-off-by: Alexey Roytman <roytman@il.ibm.com>