client: introduce model-aware explicit Conditional
The current model-based conditional is very limited in the sense that
it can only generate equality conditions.

Replace field pointers by a variadic list of Condition objects that can
express any kind of condition.

The ConditionFromModel() function will now return either an equality
conditional or an explicit one depending on the number of conditions
provided.

Signed-off-by: Adrian Moreno <amorenoz@redhat.com>