ovsdb: Improve Monitor Select Handling
As pointed out in #87 we don't handle MonitorSelect as defined
in RFC7047. If a field is nil, it should be intrepretted as true.

This commit adds differentiates between the type used for marshalling
vs. the type exposed in the API. It provides an easy way of generating
MonitorSelect objects using bools, and provides functions for
determining the value of a field. E.g `Initial()` will return
true if nil or set to true, or false if set to false.

Signed-off-by: Dave Tucker <dave@dtucker.co.uk>