notation: Correct Row notation
We had ResultRow and Row which were the same thing. The latter
placed the contents in an intermediate map[string]interface{} called
Fields while the former was just an alias for map[string]interface{}.

This commit uses an alias for map[string]interface{} everywhere for
consistency and ensures that we use this everywhere we'd expect a
Row. This allows arbitraty Row creation by casting a map directly.

Signed-off-by: Dave Tucker <dave@dtucker.co.uk>