ovsdb: Fix Mutation and Condition Marshalling
The unmarshal methods didn't use pointer receivers

Go case statements don't fallthrough

While scalar values are easily type asserted when marshalled to an
interface{} the same is not true for map, set and uuid types.
We need to use reflect to determine the OVSDB type before storing
as an interface{} value.

Signed-off-by: Dave Tucker <dave@dtucker.co.uk>