client: Refactor and replace Connect with NewOVSDBClient
This allows a client to be re-used. For example, it could be
reconnected if the underlying connection was closed.
To do this we also add an additional bool to track connection state.

It also adds SetOption which allows for an option, like the WithTLSConfig, to be
set to allow certificates to be changed.

Signed-off-by: Dave Tucker <dave@dtucker.co.uk>