ovsdb: Do not return pointer for OvsMap and OvsSet
If I recall, this was because we could return an error in those
functions. However, this means that OvsToNative and NativeToOvs are not
compatible. By returning a struct, not a pointer, everything still works
as expected AND OvsToNative and NativeToOvs become compatible without
having to write additional reflection code to deal with pointers to
structs.

Closes: #162

Signed-off-by: Dave Tucker <dave@dtucker.co.uk>