client: Speed up row access by index
This adds a new api to the cache - RowByMdoel which will calculate the
will look up by the UUID of the model, or by the matching the model
against the table indexes. This is much faster as it doesn't need to
iterate across the entire cache.

The Get API has been updated to use this new function

Fixes: #150

Signed-off-by: Dave Tucker <dave@dtucker.co.uk>