client: Fix races on access of rpcClient/connect
This fixes the locking of the rpcClient and removes the connect bool
as there was some concurrent read or write that was unprotected by mutex.
This avoids races found in the tests where the rpcClient was being used
while a disconnect was being processed.

For example, a test attempts to call `Connected` while the disconect is being
handled

Or a test makes an RPC call (e.g Transact) while a disconnect is being
handled

Locking of the client in RPC calls has been moved to right before the
call itself, with a nil check to make sure that we're still connected
before proceeding. As a result, the tests that asserted that functions
couldn't be called when disconnected were removed.

Signed-off-by: Dave Tucker <dave@dtucker.co.uk>