Makefile: Generate coverage for integration tests too
This brings back the use of `test.Short` to stop the integration tests
being run on the unit test target. The integration test target will now
generate coverage for the `client` package too. We also had to add back
the trick for combining coverage files, but at least it's in the
Makefile now.

Signed-off-by: Dave Tucker <dave@dtucker.co.uk>