cache: Disable Index Checking on Populate
The ordering of the updates from the table-update notification is
undefined. It's a hard problem to sort the update in such a way each
operation would be processed in order to ensure no indexes collided.
We can instead take the easy way out and ignore index checking
altogether as we assume the server has already take care of this.
THe cache is locked during update, so the client doesn't need to worry
about the indexes being inconsistent during the populate.

Signed-off-by: Dave Tucker <dave@dtucker.co.uk>