api: Allow use of Where for client-side ops
This uses cache.RowByCondition to enable the use of Where and variants
for operations that query the local cache. This offers increased lookup
speed when using conditions that are part of the table indexes - O(1m)
where m is the number of conditions. In cases where the field is not
part of the index, we fallback to the O(n) search.

Signed-off-by: Dave Tucker <dave@dtucker.co.uk>