api: Allow use of Where for client-side ops
This is a carry from @dave-tucker PR ovn-org/libovsdb#209.

This uses cache.RowByCondition to enable the use of Where and variants
for operations that query the local cache. This offers increased lookup
speed when using conditions that are part of the table indexes - O(1m)
where m is the number of conditions. In cases where the field is not
part of the index, we fallback to the O(n) search.

Includes additional fixes to that PR:
* Run predicate on shallow copy of cache models, which was introduced
  at some time after the original proposal of the PR.
* Removes the unnecessary model clone as the cache always gives out deep
  copies of models except for the RowsShallow method.
* List stops evaluating conditions if the result capacity is reached.

Signed-off-by: Dave Tucker <dave@dtucker.co.uk>
(cherry picked from commit 08460a1573dba9d127b2ae76fb12729e2daa0764)