client: Clean up RPC client on failed connect
This issue was caught debugging reconnect failures.
We attempt to reconnect as soon as the server disconnects, which allows
us to open a socket, but there is no reply (listdbs fails) at this
point, when we retry again, the rpcClient is not nil, so we assume we're
already connected and don't return an error.

This commit ensures that the rpcClient is set to nil if we fail to
fully connect and adjusts our tests to perform a container restart
instead of simply disconnecting the client to test the reconnect
behaviour.

Signed-off-by: Dave Tucker <dave@dtucker.co.uk>