server: copy updates for each monitor
Because each monitor filters the update and may remove rows from
the object, we need to copy it before filtering.

Specifically for a server with multiple databases, each with a
monitor, the second monitor would get an empty update because the
first monitor would have filtered out the data that didn't match
it's schema.

Signed-off-by: Dan Williams <dcbw@redhat.com>