rename DBModel to ClientDBModel
Clearly DBModel does not hold the full database model. Instead, only
the combination of model.DBModel, mapper.Mapper and ovsdb.Schema is a
useful database model as is used internally by libovsdb.

The fact that server.go had to defined a struct called DatabaseModel
with model.DBModel and ovsdb.Schema and dymanically create Mapper
objects from them is a proof of this.

In order to prepare for a DBModel refactoring, rename it to
ClientDBModel as it's the client's view of the DatabaseModel.

This patch does not contain functional changes

Signed-off-by: Adrian Moreno <amorenoz@redhat.com>