rename DBModel to DatabaseModelRequest
Clearly DBModel does not hold the full database model. Instead, only
the combination of model.DBModel, mapper.Mapper and ovsdb.Schema is a
useful database model.

The fact that server.go had to defined a struct called DatabaseModel
with model.DBModel and ovsdb.Schema and dymanically create Mapper
objects from them is a proof of this.

In order to prepare for a DBModel refactoring, rename it to
DatabaseModelRequest as it's what the client requests the DatabaseModel
to look like.

This patch does not contain functional changes

Signed-off-by: Adrian Moreno <amorenoz@redhat.com>