database: Add 2-step initialization
It is common to first create a DatabaseModel only based on the
ClientDBModel, and then add / remove the schema to it when, e.g:
when the client (re) connects.

Validation is performed when the schema is added.

Since the schema and mapper can be dynamically added and purged, protect
them with a RWMutex

Signed-off-by: Adrian Moreno <amorenoz@redhat.com>