client: Fix reconnect issues
This ensures that the cache is locked when calling Populate/Populate2
and expands the model used to represent the OVS table such that we don't
have issues caused by Monitors attempting to update unmapped fields.
In future, we should probably change the Monitor code to only ask for
ORM mapped fields.

Signed-off-by: Dave Tucker <dave@dtucker.co.uk>