Fixes synchronicity between transactions/updates
The changes to use a buffered channel to handle updates broken the
synchronization ovsdb depends on with txns. This is because OVSDB
protocol specifies an update will be sent before a txn reply. We depend
on this to ensure our cache is valid when we complete a txn.

This patch reverts the changes from
cda509cc6cbc6969df2b994f96e647eb68640c0b and while the monitor is being
setup, defers updates to a buffer. After the monitor setup is complete,
the defered updates will be processed in order. Afterwards, the db
cacheMutex is unlocked so that subsequent updates maybe processed
normally.

Signed-off-by: Tim Rozet <trozet@redhat.com>