Fixes cache population error handling
Previously failures to populate the cache would result in panic. This
should be avoided where possible and error handling should take place to
reconnect and rebuild the cache if the client and server become out of
sync. This change fixes the remaining panics to return errors, and
errors are sent to signal a reconnect and cache flush. Additionally,
errors are propagated up to the RPC handler and error responses may be
sent. Currently these responses may race with client disconnect, but at
least its a step in the right direction.

Signed-off-by: Tim Rozet <trozet@redhat.com>