Make Columns and Tables Optional
This allows for Columns to be declared optional via the
ClientDatabaseModel.

Fields may be marked as optional within the struct tag.

This allows a connection when a mapped table (and columns that reference it) aren't available at runtime to support upgrades.

To allow a user to easily see which Tables and Columns are supported two
new APIs were added.

IsTableSupported and IsColumnSupported

Signed-off-by: Dave Tucker <dave@dtucker.co.uk>