Always enforce checking all indexes
There can be a case with ovsdb-server where we receive multiple creates
in a row for an object that has different UUIDs, but the same named
index. An example of this was seen with OVN and logical router ports,
where we received a create for two of the same named ports. In this case
the create should fail, and we should reconnect and reinitialize the
cache.

On the client side, we should always validate the cache population is
legit, and if not, reconnect and try to re-init the cache.

Signed-off-by: Tim Rozet <trozet@redhat.com>