Fixes transactions during reconnect
During reconnect, transactions would block until reconnected. However,
once reconnected the "break" was only exiting the select and resulting
in an infinite loop. Additionally, there was a window of time between
Disconnect and handleDisconnectNotification where a transact operation
may be attempted and get back "error: not connected". To avoid this, add
a connected property to the client, which is more meta than if the
rpcclient is not nil and signifies the ovsdbclient is fully established.

Also added a test case to exercise the case of trying to Transact during
a reconnection and ensure it blocks appropriately.

Signed-off-by: Tim Rozet <trozet@redhat.com>