Move rpcMutex to Read Locking for Transact
Multiple in flight Transactions would be fighting over the rpcMutex lock
and effectively enforcing serialized transactions. There's no need to do
this because the rpc2 lib is being used with blocking, so it will handle
serializing transactions for us. This should help reduce contention a
little bit during periods of reconnection with high pending transaction
volumes.

Signed-off-by: Tim Rozet <trozet@redhat.com>