modelgen: add copy and equal methods for models
For each model, generate equal and copy methods that can be used by
libovsdb client for increased performance.

The approach taken leads to generating more duplicated code than what
would be strictly necessary in favor of simple and straight forward
templating that is solely based in the same type information generated
for the models.

Signed-off-by: Jaime Caamaño Ruiz <jcaamano@redhat.com>