client: don't clone objects when passing to predicate functions
This is a significant memory optimization in the case where lists of
large tables only return a few objects. Rather than deep-copying the
cache, just shallow-copy the list. Predicate functions get the uncloned
object. Then, only clone for objects that match the predicate.

This is safe over updates because updated objects are cloned before
being mutated.

Signed-off-by: Casey Callendrello <cdc@redhat.com>