api: support pointer to slice of Models for List
The whole API is based on Model defined as a pointer to a struct except
the List method which takes a pointer to a slice of structs which
technically is not equivalent to a pointer to a slice of Models. This
causes a lot of user code transforming to and from these two types.
Incidentally, this looks like it might have not been originally intended
looking at the existing bugs on the List api test.

Thus add support for the latter type and keep support of the former for
backward compatibility.

Signed-off-by: Jaime Caamaño Ruiz <jcaamano@redhat.com>