client: use LastTransactionID when the only monitor is a CondSince one
If there is only one monitor, and it's a CondSince one then we can
safely send the LastTransactionID to the server and ignore clearing
the cache if the LastTransactionID was found by the server. In all
other cases, we should pass the empty transaction ID and clear the
cache on monitor reconnect.

Signed-off-by: Dan Williams <dcbw@redhat.com>