Moves client error handling from the cache to client
In OVNK downstream we see a problem where a cache update is stalled
because it is unable to send the error to the table error channel. This
causes deadlock as well because none of the other goroutines (including
the disconnectHandler) can proceed while the cache is locked. I'm not
totally sure why the channel is blocked, according to a core dump it
looks like the send channel and the receive channel in
handleCacheErrors may be different.

However, there is an obvious race with client handlers (like
handleCacheErrors) running in goroutines and
handleDisconnectNotification. On Disconnect the stop channel is closed,
which will shutdown the handleCacheErrors goroutine. However, the cache
update may still be executing. It is not immediately shutdown while
still processing a monitor udpate. Therefore it may be unable to send an
error to the unbuffered errorChan and we could be deadlocked.

Additionally, the primary purpose of running handleCacheErrors is to
disconnect/reconnect the client in case some critical error happens. It
makes sense to move this error handling to the client rather than the
table cache. This help us remove the dependency on the cache lock. Also,
this commit adds protection to wait on disconnect for all the handlers
to shutdown before trying to reconnect.

Signed-off-by: Tim Rozet <trozet@redhat.com>