Fixes client error handling
When cache issues are detected, the client should attempt to reconnect
to rebuild the cache. However the error comparisons were incorrect and
this led to never trying to reconnect.

Additionally make missing column errors real error types, and create a
more explicit warning to the user when this type of error is detected.
Also, do not reconnect/reinit the cache on this error.

Signed-off-by: Tim Rozet <trozet@redhat.com>