Optimize RowsByCondition
RowsByCondition is heavily used in the server and can potentially be
used in the client to match conditions against the cache.

Conditions might be based on indexed fields and, for a better
performance, indexes can potentially be used to evaluate them if the
conditions test for equality.

RowsByCondition can be optimized as follows:

1. Filter out those conditions that don't test for equality.
2. With the remaining set of conditions, build all possible subsets.
3. For every subset:
   3a. Check if there is any index based on the same fields of the
       conditions of the subset. Continue to the next subset if not.
   3b. Build the index value from the values of the conditions of the
       subset and get the uuids mapped to that index value. These are
       the models that match that subset of conditions.
4. Intersect the uuids obtained in for each subset in step 3b. These
   are the models that match all those subsets.
5. Evaluate all the conditions provided to RowsByCondition against the
   models obtained in step 4. The matching models are those that match
   all conditions.

Of course, if none of the provided conditions is based on an indexed
field, then RowsByCondition needs to fallback to evaluating all the
conditions against all the rows of the table.

Signed-off-by: Jaime Caamaño Ruiz <jcaamano@redhat.com>