Benchmark predicate vs index cache lookups
As part of the effort of intricuding client indexes, one of the main
points is the performance benefit of using indexes instead of
predicates.

There is a benchmark for predicates but an incorrect implementation of
the test prevented getting correct results. Fixed it and added an index
benchmark for comparison.

Signed-off-by: Jaime Caamaño Ruiz <jcaamano@redhat.com>