Fix NativeToOVS assignment for pointer values in cache
If NativeToOVS was successful, then simply assign the value and break
out of the switch statement, but continue with the loop. Prior to this
change, only a single pointer value was assigned, and then regardless of
success or failure, ApplyModifications returned with the error code,
even if it was nil. That skipped other valid Modifications if several of
them assigned a NativeToOvs pointer value.

Signed-off-by: Andreas Karis <ak.karis@gmail.com>