Generate explicit conditionals from cache
Instead of translating model conditionals to ovsdb conditionals and
sending those to the server, perform a query on the cache through
RowsByCondition and generate uuid based conditionals off those results.
Fallback to the original conditionals if no cache hits are obtained.

This was already being done for the predicate conditional, this commit
does the same for the explicit conditional.

This has as performance benefit if RowsByCondition is able to used any
index from the conditions instead of evaluating each of them agains each
row of the table.

Signed-off-by: Jaime Caamaño Ruiz <jcaamano@redhat.com>