Generate equality conditionals from cache
Instead of generating and sending conditionals based on model indexes
and sending those to the server, perform a query on the cache with
RowsByModel and generate uuid based conditionals off those results.
Fallback to the original conditionals if no cache hits are obtained.

This was already being done for the predicate conditional, this commit
does the same for the equality conditionals.

This has two benefits:

* OVSDB server does not index schema indexes, it only ensures uniqueness
  but operations based on those indexes are not necessarily fast. By
  querying the cache for the uuids we improve the performance of index
  based operations.
* All the conditionals can fully benefit of client indexes.

Signed-off-by: Jaime Caamaño Ruiz <jcaamano@redhat.com>