client: validate transaction option
Adds a client option that enables transaction verification.

When transaction verification is enabled, transactions are serialized in
the client. This avoids the verification of a transaction when other
transactions are on flight ensuring the consistency of such
verification.

The verification is done by running the transaction in a similar way it
is done server-side but on a throw-away database built on top of the
client-cache.

The primary use case for this verification is to check for duplicate
indexes client-side instead of server-side and specifically client indexes
which can not be checked server-side. Although support for this specific
case will be introduced in later commit.

Signed-off-by: Jaime Caamaño Ruiz <jcaamano@redhat.com>