database: don't diff updates on validating transactions
For updates and mutates, the difference for each column is calculated
and set in the Modify row of an update2 notification. Calculating this
difference has a performance cost when the column is a set with
thousands of unordered values.

While calculating the Modify row is a spec requirement over the wire,
internally updates could be handled through the Old & New rows and a
transaction that is specific for validating purposes can take advantage
of this.

Make the cache capable of processing an update2 with nil Modify and
non-nil Old & New as if it were a normal update notification and avoid
setting Modify for updates on validating transactions.

Signed-off-by: Jaime Caamaño Ruiz <jcaamano@redhat.com>