Fix client handling for modify for optional types
When receiving an update for a column who contains a optional set type,
and the updated value is the same as what previously existed, the
current behavior may panic when checking length of the pointer. We
should never receive an update with equal values, but in case we do
log an error and move on (since the outcome is the same value anyway).

Also, we should never receive a row modification with more than 1
value. In that case return an error.

Signed-off-by: Tim Rozet <trozet@redhat.com>