move database & transaction to its own package
Adding constraint validation functionality to the transaction will
increase it's code base and cause further changes to the database
interface and implementation, given them more entity relative to the
actual server implementation. For better organization and clarity, it
makes sense to move them to its own package.

As a consequnce, move mutate to ovsdb package and refactor some test
utilities so that they can be reused accross the new database package
and the server package.

Signed-off-by: Jaime Caamaño Ruiz <jcaamano@redhat.com>