cache: fix index exists checks
The cache might be used as part of a validating transaction so an
schema index can temporarily have multiple UUIDs for an index as models
are manipulated with index check disabled.

So it is not enought to check if a particualr UUID is not associated to
an index, we need to actually check that it is the only UUID associated
to the index.

Strengthen the exists check overall.

Signed-off-by: Jaime Caamaño Ruiz <jcaamano@redhat.com>