cache: don't update indexes on error
When there was an error processing an index, previous indexes would have
been updated as the uuidset methods update in-place for performance.

This error though is very unlikely to happen, it would require a schema
mismatch on an indexed field, or using sets as a client index (which can't
happen because this is validated).

Signed-off-by: Jaime Caamaño Ruiz <jcaamano@redhat.com>