Split database package
The main purpose of this change is to have a package for common database
types and interfaces that can be used throughout the project without
incurring in circular dependencies.

To that end, the in-memory database implementation is moved to its own
subpackage.

The transaction implementation is also moved to its own subpackage given
that it can work with any database implementing the Database interface.

At the interface level though it should be expected that a transaction
implementation could have a more tighter underlying coupling with a
database implementation. So it makes sense to introduce a transaction
interface, and for transactions to be provided by database
implementations rather than being built directly by the server.

Signed-off-by: Jaime Caamaño Ruiz <jcaamano@redhat.com>