Add inactivity probe for ovsdb connection
This enhances ovsdb client to sends echo request periodically at specified
interval to ovsdb server, when echo request fails consecutively 2 * interval
then consider connection as inactive and then attempts to reconnect with server.
This ensures early detection of ovsdb server connectivity issues and reconnect
to the server immediately when it's back to operational.

Signed-off-by: Periyasamy Palanisamy <pepalani@redhat.com>