Reuse reconnect logic for inactivity check
This reuses reconnect code associated with handling disconnect notification
for the echo failure, It makes code more readble and avoids unneccessary use
of additional mutexes and flags.

It also avoids sending echo unnecessarily when ovsdb traffic going on with
server connection and that's being for connection aliveness check which
saves some cpu cycles and echo traffic on the wire.

Signed-off-by: Periyasamy Palanisamy <pepalani@redhat.com>