ovsdb: partial revert of "mapper: Add _uuid when calling NewRow"
This is a partial revert of commit bbcbf5a93ea866ec16ec58886c636d6daa887239

The reverted commit prevents marshalling named UUIDs into a Row's _uuid
field. But that's kind of silly because the _uuid field shouldn't ever be
filled in for requests from client -> server, whether named or real.

There are only two operations that use Rows: Insert and Update. On Insert
ovsdb-server overwrites the given Row's _uuid with either the UUID given
in the Operation, or a generated one. On Update it throws an error since
_uuid is immutable. So there's zero point to letting _uuid through.

But it may be useful internally when dealing with Rows to allow both
regular and named UUIDs in the _uuid field, whether for testcases or
constructing strings of Operations before sending to the server.

Thus, do the same thing in Insert that Update already does: specifically
remove _uuid from the row as late as possible, right before sending the
Operation to the server via RPC.

Signed-off-by: Dan Williams <dcbw@redhat.com>