updates: fix merge resulting in empty optional values
An optional atomic value is cleared with an empty set instead of a nil
set. Thus use an empty set as an original empty value so that when an op
that sets a value is merged with an op that clears the value, the
comparison with an empty orignal value results in a noop.

Signed-off-by: Jaime Caamaño Ruiz <jcaamano@redhat.com>