Remove references from test schema
Once we introduce referential integrity, existing tests would start to
fail as they are not accounting for it. So its better to remove the
references from the test schema which is something that current tests
are not leveraging, and then expand the schema with other references
that the referenctial integrity tests would leverage.

Signed-off-by: Jaime Caamaño Ruiz <jcaamano@redhat.com>