transaction: fix results on error
The number of results of a transaction should always have a minimum
length of the number of operations passed to it even if there is an
error. The recent introduction of the named UUID expansion in the
transaction broke this requirement. Changed the error handling approach
so that this is less likely to happen in the future.

Also removed the panic when initializing the transaction cache.

Signed-off-by: Jaime Caamaño Ruiz <jcaamano@redhat.com>