server: implement referential integrity
Implement referential integrity and garbage collection as described in
RFC7047.

This is achieved keeping track of inverse references. A reference
tracker is used in the transaction to keep these references updated, to
check for referential integrity related violations and to add additional
updates to the transaction resulting from reference garbage collection.

The inverse references are stored in the database on commit.

The updates resulting from reference garbage collection are also sent to
any monitoring clients as expected.

Signed-off-by: Jaime Caamaño Ruiz <jcaamano@redhat.com>